moving stuff out of models into config
